Feature/#18/add graph ql server basics

+++ db/migrate/20190425083149_create_prices.rb
@@ -2,7 +2,7 @@ class CreatePrices < ActiveRecord::Migration[5.2]
   def change
     create_table :prices do |t|
       t.string :name
-      t.integer :price
+      t.float :price
       t.boolean :group_price
       t.integer :age_from
       t.integer :age_to
